Removes automated stress test that only runs under Node v18
