+++ .github/workflows/publish.yml
@@ -29,7 +29,6 @@ jobs:
           node-version: ${{ matrix.node-version }}
       - run: npm ci
       - run: npm test
-#      - run: npm run stress-test-auto
 
   Conditional-Publish-to-NPM:
     needs: [Lint, Automated-Tests]
