+++ spec/stress/rapid_requests_node14_spec.js
@@ -134,7 +134,7 @@ describe('Rapid requests', function () {
   /* This is a functional test, that the server behaves correctly when overloaded.
    * It must pass on every machine. If no 429s are returned, increase `num` to make the test harder. */
   it('returns "429 Too Many Requests" when a burst of puts or gets continues too long', async function () {
-    const delayMs = 2; const num = 3000;
+    const delayMs = 1; const num = 3000;
     const puts = []; const gets = [];
     for (let i = 0; i < num; ++i) {
       const data = 'ABC' + String(1000 + i);
